Removing excess code in main css file
This commit focused on removing most of the code from the main css file, as majority of it was related to the other html pages that were deleted, as well as content that was deleted from the index page as well. I also used this commit to begin incorporating my branding colours.