Experimenting with attaching images
This commit focused on attaching my png image file, and to see how it interacts with my website. An error has come up, so I will be addressing that. I also tweaked the side bar section. It is a very large section of the code, however I decided to keep most if it, as I find that the effects give a professional polish to my website, and adds to the interactive user experience.