Clean up css file, section one
For this commit, I focused on cleaning up my code further, as I discovered that there was more code that did not need to be present. I also fixed the section setup so I could start adding text for hyperlinks. I do plan to change the colours, as I want to only have colours that are a part of my branding.