Start adding sections and adding hyperlinks
This commit focused on myself figuring how to incorporate my png images, as well as finally ading my hyperlinks. I did alot of experimentation, especially between github and photoshop, as I wanted all my text to be as aligned as possible, as well as being mindful of how I placed text so that there were not large gaps between texts.